+++ cyclops-ui/src/components/shared/HelmReleaseDetails/custom.css
@@ -1,0 +1,12 @@
+.ant-col.ant-form-item-label {
+  padding: 0 !important;
+}
+
+label.ant-form-item-required:before {
+  bottom: 1em !important;
+  position: relative !important;
+}
+
+code {
+  color: #000080;
+}
