+++ cyclops-ui/src/components/shared/HelmReleaseDetails/index.tsx
@@ -1,0 +1,1 @@
+export { HelmReleaseDetails } from "./HelmReleaseDetails";
