+++ cyclops-ui/src/components/shared/index.tsx
@@ -1,1 +1,2 @@
 export { ModuleResourceDetails } from "./ModuleResourceDetails";
+export { HelmReleaseDetails } from "./HelmReleaseDetails";
