+++ cyclops-ui/tsconfig.json
@@ -1,30 +1,3 @@
-// {
-//   "compilerOptions": {
-//     "target": "es5",
-//     "lib": ["dom", "dom.iterable", "esnext"],
-//     "allowJs": true,
-//     "skipLibCheck": true,
-//     "esModuleInterop": true,
-//     "strict": true,
-//     "forceConsistentCasingInFileNames": true,
-//     "noFallthroughCasesInSwitch": true,
-//     "resolveJsonModule": true,
-//     "isolatedModules": true,
-//     "noEmit": true,
-
-//     "jsx": "react-jsx",
-//     "module": "ESNext",
-//     "declaration": true,
-//     "declarationDir": "types",
-//     "sourceMap": true,
-//     "outDir": "dist",
-//     "moduleResolution": "node",
-//     "allowSyntheticDefaultImports": true,
-//     "emitDeclarationOnly": true
-//   },
-//   "include": ["src/", "src/types"]
-// }
-
 {
   "compilerOptions": {
     "target": "ESNext", // Adjust the target version if needed
